removed all the casts to const ints

+++ src/p_ceilng.cpp
@@ -28,10 +28,10 @@ void T_MoveCeiling (ceiling_t* ceiling)
 	
     switch(ceiling->direction)
     {
-	case static_cast<const int>(Direction::WAITING):
+	case Direction::WAITING:
 	// IN STASIS
 	break;
-	case static_cast<const int>(Direction::UP):
+	case Direction::UP:
 	// UP
 	res = T_MovePlane(ceiling->sector,
 			  ceiling->speed,
@@ -75,7 +75,7 @@ void T_MoveCeiling (ceiling_t* ceiling)
 	}
 	break;
 	
-	case static_cast<const int>(Direction::DOWN):
+	case Direction::DOWN:
 	// DOWN
 	res = T_MovePlane(ceiling->sector,
 			  ceiling->speed,
